requirements.txt
pandas
numpy
matplotlib
seaborn
jupyter
google-cloud-bigquery
google-cloud-storage
streamlit
folium
great-expectations
pyarrow
requests
db-dtypes
kafka-python